requirements.txt
kafka-python==2.0.2
aiokafka==0.7.2
schedule==1.1.0
Flask==3.0.2
Werkzeug==3.0.1
Faker==15.1.3
requests==2.28.1
pyyaml==6.0.1
python-dotenv==0.21.0
flask-cors==4.0.0
aiohttp==3.9.3